+++ ast/docs.go
@@ -12,12 +12,12 @@ func MakeDocs(comments []string) *Docs {
 	docsStrings := make([]string, 0)
 	for _, comment := range comments {
 		if strings.HasPrefix(comment, "///") {
-			docsStrings = append(docsStrings, strings.TrimSpace(strings.TrimPrefix(comment, "///")))
+			docsStrings = append(docsStrings, strings.TrimPrefix(strings.TrimPrefix(strings.TrimSpace(comment), "///"), " "))
 		} else if strings.HasPrefix(comment, "/**") {
 			trimmed := strings.TrimSpace(strings.TrimSuffix(strings.TrimPrefix(comment, "/**"), "*/"))
 			lines := strings.Split(trimmed, "\n")
 			for _, line := range lines {
-				docsStrings = append(docsStrings, strings.TrimSpace(strings.TrimPrefix(strings.TrimSpace(line), "*")))
+				docsStrings = append(docsStrings, strings.TrimPrefix(strings.TrimPrefix(strings.TrimSpace(line), "*"), " "))
 			}
 		}
 	}